【132_trunk】sync

Created-by: belong326
Commit-by: belong326
Merged-by: ohci1
Description: ### 一、内容说明（相关的Issue）



### 二、建议测试周期和提测地址  
  建议测试完成时间：xxxx.xx.xx  
  投产上线时间：xxxx.xx.xx  
  提测地址：CI环境/压测环境  
  测试账号：  

### 三、变更内容
  * 3.1 关联PR列表

  * 3.2 数据库和部署说明  
    1. 常规更新 
    2. 重启unicorn
    3. 重启sidekiq
    4. 迁移任务：是否有迁移任务，没有写 "无"
    5. rake脚本：`bundle exec xxx RAILS_ENV = production`；没有写 "无"

  * 3.4 其他技术优化内容（做了什么，变更了什么）
    - 重构了 xxxx 代码
    - xxxx 算法优化


  * 3.5 废弃通知（什么字段、方法弃用？）



  * 3.6  后向不兼容变更（是否有无法向后兼容的变更？）


  
### 四、研发自测点（自测哪些？冒烟用例全部自测？）
  自测测试结论：


### 五、测试关注点（需要提醒QA重点关注的、可能会忽略的地方）
  检查点：

| 需求名称 | 是否影响xx公共模块 | 是否需要xx功能 | 需求升级是否依赖其他子产品 |
|------|------------|----------|---------------|
| xxx  | 否          | 需要       | 不需要           |
|      |            |          |               |

  接口测试：

  性能测试：

  并发测试：

  其他：



See merge request: openharmony-tpc/chromium_src!3394

Revert commit: merge 132_trunk into 132_trunk
This revert commit: d87e4c45c11415fd0008bfb1973684182f693ce6

+++ arkweb/ohos_nweb/src/cef_delegate/nweb_history_list_impl.cc
@@ -90,7 +90,7 @@ int32_t NWebHistoryListImpl::GetCurrentIndex() {
 }
 
 std::shared_ptr<NWebHistoryItem> NWebHistoryListImpl::GetItem(int32_t index) {
-  if (index < 0 || index >= static_cast<int32_t>(item_list_.size())) {
+  if (index < 0 || index >= item_list_.size()) {
     return nullptr;
   }
   return item_list_[index];
